% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ottawa_roads_shp}
\alias{ottawa_roads_shp}
\title{2021 Statistics Canada Road Network for Ottawa, Ontario}
\format{
A data frame with class sf in CRS NAD/MTM zone 9 (32189) and 33,983 rows and 5 variables:
\describe{
\item{NGD_UID}{Unique road segment identifier}
\item{NAME}{Road segment name}
\item{RANK}{Road rank, lower numbers generally mean bigger/faster roads,
   \url{https://www12.statcan.gc.ca/census-recensement/2021/geo/ref/domain-domaine/index2021-eng.cfm?lang=e&id=RANK}}
\item{CLASS}{Road class, lower numbers generally but do not always mean bigger/faster roads,
   \url{https://www12.statcan.gc.ca/census-recensement/2021/geo/ref/domain-domaine/index2021-eng.cfm?lang=e&id=CLASS}}
\item{geometry}{LINESTRING defining road segment geometry}
}

This data is licensed under the Statistics Canada Open Data License
(\url{https://www.statcan.gc.ca/en/reference/licence}).
Adapted from Statistics Canada, 2021 Census Road Network File, 2022-09-28.
This does not constitute an endorsement by Statistics Canada of this product.
}
\usage{
ottawa_roads_shp
}
\description{
Spatial dataset for road networks in Ottawa, Ontario, provided by Statistics
Canada, \url{https://www12.statcan.gc.ca/census-recensement/2021/geo/sip-pis/rnf-frr/index2021-eng.cfm?year=21}.
}
\keyword{datasets}
