% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{region_shp}
\alias{region_shp}
\title{Synthetic region shapefile for testing}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 3 columns.
}
\usage{
region_shp
}
\description{
A small shapefile for testing.
}
\keyword{datasets}
