% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{road_shp}
\alias{road_shp}
\title{Synthetic road shapefile for testing}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 2 columns.
}
\usage{
road_shp
}
\description{
A small shapefile for testing.
}
\keyword{datasets}
